+++ expense-service/src/main/java/com/spendsmart/expense/repository/ExpenseRepository.java
@@ -11,40 +11,40 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
+public interface ExpenseRepository extends JpaRepository<Expense, Long> {
 
     // By userId
-    List<Expense> findByUserId(int userId);
+    List<Expense> findByUserId(Long userId);
 
     // By userId + type (EXPENSE / SPLIT)
-    List<Expense> findByUserIdAndType(int userId, String type);
+    List<Expense> findByUserIdAndType(Long userId, String type);
 
     // By categoryId
-    List<Expense> findByCategoryId(int categoryId);
+    List<Expense> findByCategoryId(Long categoryId);
 
     // By userId + exact date
-    List<Expense> findByUserIdAndDate(int userId, LocalDate date);
+    List<Expense> findByUserIdAndDate(Long userId, LocalDate date);
 
     // By userId + date range
-    List<Expense> findByUserIdAndDateBetween(int userId, LocalDate start, LocalDate end);
+    List<Expense> findByUserIdAndDateBetween(Long userId, LocalDate start, LocalDate end);
 
     // Sum of all amounts for a user
     @Query("SELECT COALESCE(SUM(e.amount), 0) FROM Expense e WHERE e.userId = :userId")
-    Double sumAmountByUserId(@Param("userId") int userId);
+    Double sumAmountByUserId(@Param("userId") Long userId);
 
     // Sum by userId + categoryId
     @Query("SELECT COALESCE(SUM(e.amount), 0) FROM Expense e WHERE e.userId = :userId AND e.categoryId = :categoryId")
-    Double sumAmountByUserIdAndCategoryId(@Param("userId") int userId, @Param("categoryId") int categoryId);
+    Double sumAmountByUserIdAndCategoryId(@Param("userId") Long userId, @Param("categoryId") Long categoryId);
 
     // By expenseId
-    Optional<Expense> findByExpenseId(int expenseId);
+    Optional<Expense> findByExpenseId(Long expenseId);
 
     // Delete by expenseId
-    void deleteByExpenseId(int expenseId);
+    void deleteByExpenseId(Long expenseId);
 
     // By month + year + userId
     @Query("SELECT e FROM Expense e WHERE e.userId = :userId AND MONTH(e.date) = :month AND YEAR(e.date) = :year")
-    List<Expense> findByUserIdAndMonthAndYear(@Param("userId") int userId,
+    List<Expense> findByUserIdAndMonthAndYear(@Param("userId") Long userId,
                                                @Param("month") int month,
                                                @Param("year") int year);
 
@@ -52,8 +52,8 @@ public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
     @Query("SELECT e FROM Expense e WHERE e.userId = :userId AND " +
            "(LOWER(e.title) LIKE LOWER(CONCAT('%', :keyword, '%')) OR " +
            "LOWER(e.notes) LIKE LOWER(CONCAT('%', :keyword, '%')))")
-    List<Expense> searchByKeyword(@Param("userId") int userId, @Param("keyword") String keyword);
+    List<Expense> searchByKeyword(@Param("userId") Long userId, @Param("keyword") String keyword);
 
     // By userId + categoryId
-    List<Expense> findByUserIdAndCategoryId(int userId, int categoryId);
+    List<Expense> findByUserIdAndCategoryId(Long userId, Long categoryId);
 }
